Change ingredients components structure.
