+++ src/components/ingredients-section/ingredients-section.jsx
@@ -3,25 +3,15 @@ import { useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch } from 'react-redux';
 import { menuItemPropTypes } from '../../utils/constants';
+import BurgerIngredient from '../burger-ingredient/burger-ingredient';
 import ingredientsSectionStyles from './ingredients-section.module.css';
-import { CurrencyIcon } from '@ya.praktikum/react-developer-burger-ui-components';
-import { OPEN_MODAL, OPEN_INGREDIENT_MODAL, SET_CURRENT_INGREDIENT_DETAILS, SET_TAB_OFFSETTOP } from '../../services/actions';
+import { SET_TAB_OFFSETTOP } from '../../services/actions';
 
 function IngredientsSection(props) {
     let minOffsetTop = 0;
     const dispatch = useDispatch();
     const sectionElement = useRef(null);
 
-    const handleClick = (e) => {
-        let currentTargetId = e.currentTarget.id;
-        dispatch({ type: OPEN_MODAL });
-        dispatch({ type: OPEN_INGREDIENT_MODAL });
-        dispatch({
-            type: SET_CURRENT_INGREDIENT_DETAILS,
-            id: currentTargetId
-        });
-    }
-    
     useEffect(() => {
         let tabName = sectionElement.current.getAttribute("data-tabname");
         if (tabName === "one") {
@@ -41,17 +31,7 @@ function IngredientsSection(props) {
             <ul className={`pl-4 ${ingredientsSectionStyles.ingredientsContainer}`}>
                 {
                     props.data.map((item) => {
-                        return (item.type == props.ingredientsType) &&
-                            <li className={`mb-8 ${ingredientsSectionStyles.ingredient}`} key={item._id} onClick={handleClick} id={item._id}>
-                                <img className="pl-4 pr-4" src={item.image} />
-                                <p className="pt-1 pb-1 flexContainerJcCenter">
-                                    <span className={`pr-1 text text_type_digits-default ${ingredientsSectionStyles.price}`}>{item.price}</span>
-                                    <CurrencyIcon type="primary" />
-                                </p>
-                                <p className={`text text_type_main-default ${ingredientsSectionStyles.ingredientsName}`}>
-                                    {item.name}
-                                </p>
-                            </li>
+                        return <BurgerIngredient item={item} data={props.data} key={item._id} ingredientsType={props.ingredientsType} />;
                     })
                 }
             </ul>
